requirements.txt
click==6.6
python-slugify==1.2.0
